requirements.txt
python-dotenv~=1.0
uvicorn~=0.27
fastapi~=0.109
open_clip_torch~=2.24
sentence-transformers~=3.0